+++ qg-cms-front/lib/NavLinks.js
@@ -1,8 +1,0 @@
-const NavLinks = [
-    { href: '/', title: 'Home' },
-    { href: '/articles/posts', title: 'Articles' },
-    { href: 'https://qgriffith.me/resume', title: 'Resume' },
-    { href: '/about', title: 'About' },
-  ]
-  
-  export default NavLinks