+++ qg-cms-front/pages/404.js
@@ -9,9 +9,12 @@ export default function Custom404() {
     const global = await queryAPI('/global', {
       populate: "*"
     })
+    const navbar = await queryAPI('/navbars')
+
     return {
         props: {
-            global: global.data
+            global: global.data,
+            navbar: navbar.data
         },
         revalidate: 500
     }