+++ qg-cms-front/pages/500.js
@@ -8,9 +8,12 @@ export default function Custom500() {
     const global = await queryAPI('/global', {
       populate: "*"
     })
+    const navbar = await queryAPI('/navbars')
+
     return {
         props: {
-            global: global.data
+            global: global.data,
+            navbar: navbar.data
         },
         revalidate: 500
     }