+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2022-8-6
+### Added
+- Pull NavBar data from Strapi instead of a hard coded file
+- Add Strava Icont to the about page
+
+### Changed
+- Increased the size of the logo to better match the increased navbar
+- Increased the size of the site title
+
 ## [1.1.2] - 2022-8-5
 ### Added
 - Installed the sharp module to speed up images